+++ sema4ai/codegen/commands.py
@@ -1042,6 +1042,13 @@ COMMANDS = [
         hide_from_command_palette=True,
         icon="$(go-to-file)",
     ),
+    Command(
+        "sema4ai.dropAllDataSources",
+        "Remove All Data Sources",
+        server_handled=False,
+        add_to_package_json=True,
+        icon="$(trash)",
+    ),
 ]